+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -7,54 +7,54 @@ public class StatsServiceTest {
     @Test//сумма всех продаж
     public void testSumAllSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 180;
-        int actual = service.sumAllSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 180;
+        long actual = service.sumAllSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test//средняя сумма продаж в месяц
     public void averageSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 15;
-        int actual = service.averageSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 15;
+        long actual = service.averageSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test//номер месяца, в котором был пик продаж
     public void maxSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 8;
-        int actual = service.maxSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 8;
+        long actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test//номер месяца, в котором был минимум продаж
     public void minSales() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 9;
-        int actual = service.minSales(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 9;
+        long actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test//количество месяцев, в которых продажи были ниже среднего
     public void salesBelowAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 5;
-        int actual = service.salesBelowAverage(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 5;
+        long actual = service.salesBelowAverage(sales);
         Assertions.assertEquals(expected, actual);
     }
 
     @Test//количество месяцев, в которых продажи были выше среднего
     public void salesAboveAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expected = 5;
-        int actual = service.salesAboveAverage(sales);
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long expected = 5;
+        long actual = service.salesAboveAverage(sales);
         Assertions.assertEquals(expected, actual);
     }
 }
